+++ termsaver/__init__.py
@@ -47,9 +47,12 @@ import errno
 # Python built-in modules
 #
 import sys
+from signal import SIGINT, signal
 
 from termsaver.termsaverlib import common, constants, exception
 from termsaver.termsaverlib.helper.smartformatter import SmartFormatter
+from termsaver.termsaverlib.helper.utilities import (hide_stdout_cursor,
+                                                     show_stdout_cursor)
 from termsaver.termsaverlib.i18n import _
 #
 # Internal modules
@@ -87,6 +90,11 @@ Options:
  -h, --help     Displays this help message
  -v, --verbose  Displays python exception errors (for debugging)
 
+Enhanced Features:
+ * Install the following modules to enable enhanced features:
+    * pynput - Enables the 'Press any key to exit' feature.
+    * pygments - Colorizes the output of the Programmer screen.
+
 Refer also to each screen's help by typing: %(app_name)s [screen] -h
 """) % {
         'app_name': constants.App.NAME,
@@ -100,8 +108,15 @@ Refer also to each screen's help by typing: %(app_name)s [screen] -h
 def skip(arg = None):
     pass
 
+def handler(signal_received, frame):
+    # Handle any cleanup here
+    print('SIGINT or CTRL-C detected. Exiting gracefully')
+    show_stdout_cursor()
+    sys.exit(0)
             
 def entryPoint():
+    signal(SIGINT, handler)
+    hide_stdout_cursor()
     tscreen = getScreen()
     if tscreen:
         (screen, parser) = tscreen
@@ -137,6 +152,7 @@ def getScreen():
         verbose = True if args.verbose else False
         if args.screen == None or (args.screen == None and args.h0elp == True):
             usage()
+            show_stdout_cursor()
             sys.exit(0)
 
         # Find the screen we're using and create it's parser as well as check validity with args.
@@ -152,6 +168,7 @@ def getScreen():
         if screen == None:
             print(_("Invalid Screen."))
             usage()
+            show_stdout_cursor()
             sys.exit(0)
 
         # Pass the parser to the selected screen.
@@ -173,6 +190,7 @@ def getScreen():
             common.prettify_exception(e)
 
         # Just finish gracefully
+        show_stdout_cursor()
         sys.exit(0)
 
     except exception.TermSaverException as e:
@@ -262,6 +280,7 @@ def getScreen():
         if e.help_msg not in (None, ''):
             print(e.help_msg)
 
+        show_stdout_cursor()
         sys.exit(error_number)
 
     except Exception as e:
@@ -293,7 +312,7 @@ option --verbose and copy the output when you are filling
 the bug report, that will help track faster the problem.
 Thanks!
 """))
-
+        show_stdout_cursor()
         sys.exit(errno.EPERM)
 if __name__ == '__main__':
     #
