+++ termsaver/termsaverlib/screen/asciiartfarts.py
@@ -34,14 +34,17 @@ The screen class available here is:
     * `AsciiArtFartsScreen`
 """
 
+import time
+
 from termsaver.termsaverlib.i18n import _
 #
 # Internal modules
 #
-from termsaver.termsaverlib.screen.base.rssfeed import SimpleRSSFeedScreenBase
+from termsaver.termsaverlib.screen.base.urlfetcher import SimpleUrlFetcherBase
+from termsaver.termsaverlib.screen.helper.position import PositionHelperBase
 
 
-class AsciiArtFartsScreen(SimpleRSSFeedScreenBase):
+class AsciiArtFartsScreen(SimpleUrlFetcherBase, PositionHelperBase):
     """
     A simple screen that connects to asciiartfarts.com RSS feeds to display
     ascii arts on screen.
@@ -77,25 +80,28 @@ class AsciiArtFartsScreen(SimpleRSSFeedScreenBase):
 
         NOTE: Maybe NSFW (Not Safe For Work)
         """
-        SimpleRSSFeedScreenBase.__init__(self,
+        SimpleUrlFetcherBase.__init__(self,
             "asciiartfarts",
             _("displays ascii images from asciiartfarts.com (NSFW)"),
             parser,
-            'http://www.asciiartfarts.com/farts.rss',
-            ["description"],
-            "%(description)s",
+            'http://www.asciiartfarts.com/random.cgi',
         )
 
         if self.parser:
           self.parser.add_argument("-d", "--delay", help="Delay in seconds between images", type=int, default=5)
 
-        # set defaults for this screen
-        self.cleanup_per_cycle = True
-        self.sleep_between_items = 5
-        self.center_vertically = True
-        self.center_horizontally = True
-        self.clean_dirt = ['<pre>', '</pre>']
-        self.cleanup_per_item = True
+    def _run_cycle(self):
+        """
+        Executes a cycle of this screen. Overriden from its superclass because
+        it needs to must randomize the URL to be fetched in every screen cycle.
+        """
+        self.clear_screen()
+        new_text = self.process_data(self.fetch(self.url)).decode("utf-8")
+        self.get_terminal_size()
+        new_text = self.center_text_vertically(new_text)
+        new_text = self.center_text_horizontally(new_text)
+        self.typing_print(new_text)
+        time.sleep(self.sleep_between_items)
 
     def _parse_args(self, launchScreenImmediately=True):
       args, unknown = self.parser.parse_known_args()
@@ -104,3 +110,8 @@ class AsciiArtFartsScreen(SimpleRSSFeedScreenBase):
         self.autorun()
       else:
         return self
+
+    def process_data(self, data):
+      data_string = data.decode('utf-8')
+      data_string = data_string.split('<pre>')[2].split('</pre>')[0]
+      return data_string.encode('utf-8')
