+++ termsaver/termsaverlib/screen/base/__init__.py
@@ -79,19 +79,25 @@ consistent).
 
 """
 
-import argparse
-import getopt
 #
 # Python built-in modules
 #
-import os
+import subprocess
 import sys
+
+from termsaver.termsaverlib.helper.utilities import show_stdout_cursor
+
+pynput_installed = None
+reqs = subprocess.check_output([sys.executable, '-m', 'pip', 'freeze'])
+installed_packages = [r.decode().split('==')[0] for r in reqs.split()]
+if 'pynput' in installed_packages:
+    pynput_installed = True
+    from pynput import keyboard
 
 #
 # Internal modules
 #
-from termsaver.termsaverlib import common, constants, exception
-from termsaver.termsaverlib.helper.smartformatter import SmartFormatter
+from termsaver.termsaverlib import constants
 from termsaver.termsaverlib.i18n import _
 from termsaver.termsaverlib.screen.helper import ScreenHelperBase
 
@@ -177,6 +183,26 @@ class ScreenBase(ScreenHelperBase):
 
         self.name = name
         self.description = description
+        
+        if pynput_installed is not None:
+            self.listener = keyboard.Listener(
+                on_press=self.on_press,
+                on_release=self.on_release
+            )
+    
+    def on_press(self, key):
+        """
+        This method is called when a key is pressed.
+        """
+        if pynput_installed is not None:
+            self.listener.stop()
+    
+    def on_release(self, key):
+        """
+        This method is called when a key is released.
+        Unused for now, but leaving it in so we have options in the future.
+        """
+        pass
 
     def autorun(self, loop=True):
         """
@@ -202,7 +228,13 @@ class ScreenBase(ScreenHelperBase):
         # execute the cycle
         self.clear_screen()
 
-        while(loop):
+        if pynput_installed is not None:
+            self.listener.start()
+        while(
+            (loop and pynput_installed is None)
+            or
+            (loop and pynput_installed is not None and self.listener.is_alive())
+        ):
             try:
                 self._run_cycle()
             except KeyboardInterrupt as e:
@@ -215,6 +247,7 @@ class ScreenBase(ScreenHelperBase):
             # Clear screen if appropriate
             if self.cleanup_per_cycle:
                 self.clear_screen()
+        show_stdout_cursor()
 
     def _run_cycle(self):
         """
